+++ crates/affinidi-did-resolver/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## DID webvh method (0.1.5)
 
+* **CHANGE:** `DIDWebVHState` is now the entry point to working with webvh
+LogEntries and WitnessProofs
 * **FEATURE:** Witness Proofs are now optimized based on webvh 1.0 spec
   * Prior witness entries will be removed, only keeping the latest witness proof
   for each witness